change list measure to  mesures and fixed rm commands
