+++ src/ml/commands/dataset.py
@@ -3,7 +3,8 @@ import os
 from ml.utils.config import get_settings
 from ml.utils.order import order_table_print
 from ml.utils.numeric_functions import humanize_bytesize
-from ml.clf.wrappers import DataDrive, ListMeasure
+from ml.clf.wrappers import DataDrive
+from ml.clf.measures import ListMeasure
 from ml.ds import DataSetBuilder, DataLabel
 from ml.utils.files import rm, get_models_path, delete_file_model
 from ml.utils.files import get_date_from_file, get_models_from_dataset
@@ -19,19 +20,13 @@ def run(args):
     elif args.rm:
         try:
             dataset = DataSetBuilder(args.rm)
-            models_path = get_models_from_dataset(dataset.md5(), settings["checkpoints_path"])
             print("Dataset: {}".format(dataset.url()))
             rm(dataset.url())
-            for model_path, _ in models_path:
-                print("Model: {}".format(model_path))
-                rm(model_path)
         except IOError:
             pass
-        delete_orphans(settings["checkpoints_path"])
         print("Done.")
-    elif args.clean:
-        delete_orphans(settings["checkpoints_path"])
-        print("Done.")
+    #elif args.clean:
+    #    print("Done.")
     elif args.used_in:
         dataset = DataSetBuilder(args.used_in)
         for _, model_path_meta in get_models_from_dataset(dataset.md5(), settings["checkpoints_path"]):
@@ -78,13 +73,3 @@ def get_model_dataset(classes):
                 "md5", os.path.join(
                     settings["checkpoints_path"], clf, name_version, name_version))
             yield clf, name_version, md5, dataset_md5.get(md5, None)
-
-
-def delete_orphans(checkpoints_path):
-    for clf, model_name_v, md5, dataset_name in get_model_dataset(get_models_path(checkpoints_path)):
-        if dataset_name is None:
-            name_fragments = model_name_v.split(".")
-            model_name = ".".join(name_fragments[:-1])
-            model_version = name_fragments[-1]
-            path = delete_file_model(clf, model_name, model_version, checkpoints_path)
-            print(path)
